resolve circle

+++ src/main/java/com/a6raywa1cher/hackservspring/service/impl/UserServiceImpl.java
@@ -6,6 +6,7 @@ import com.a6raywa1cher.hackservspring.model.UserRole;
 import com.a6raywa1cher.hackservspring.model.VendorId;
 import com.a6raywa1cher.hackservspring.model.repo.UserRepository;
 import com.a6raywa1cher.hackservspring.security.jwt.service.RefreshTokenService;
+import com.a6raywa1cher.hackservspring.service.TeamService;
 import com.a6raywa1cher.hackservspring.service.UserService;
 import com.a6raywa1cher.hackservspring.service.dto.UserInfo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,15 +22,15 @@ import java.util.stream.StreamSupport;
 
 @Service
 public class UserServiceImpl implements UserService {
-	//private final TeamService teamService;
+	private final TeamService teamService;
 	private final UserRepository repository;
 	private final PasswordEncoder passwordEncoder;
 	private final RefreshTokenService refreshTokenService;
 
 	@Autowired
-	public UserServiceImpl(UserRepository repository, PasswordEncoder passwordEncoder,
+	public UserServiceImpl(TeamService teamService, UserRepository repository, PasswordEncoder passwordEncoder,
 						   RefreshTokenService refreshTokenService) {
-		//this.teamService = teamService;
+		this.teamService = teamService;
 		this.repository = repository;
 		this.passwordEncoder = passwordEncoder;
 		this.refreshTokenService = refreshTokenService;
@@ -158,12 +159,12 @@ public class UserServiceImpl implements UserService {
 	@Transactional(rollbackOn = Exception.class)
 	public void deleteUser(User user) {
 		refreshTokenService.invalidateAll(user);
-		/*if(user.getTeam() != null){
+		if (user.getTeam() != null) {
 			teamService.deleteMember(user.getTeam(), user);
 		}
-		if(teamService.getTeamRequestForUser(user).isPresent()){
-			teamService.deleteRequest(teamService.getTeamRequestForUser(user).get(),user);
-		}*/
+		if (teamService.getTeamRequestForUser(user).isPresent()) {
+			teamService.deleteRequest(teamService.getTeamRequestForUser(user).get(), user);
+		}
 		repository.delete(user);
 	}
 }
